Conflicts:
	app/forms.py
	app/views.py

+++ app/controllers/familyBackground.py
@@ -15,19 +15,19 @@ class familyBackground():
 	def insertFamilyBackground(self,ci_user,asthma,cancer,heartdisease,diabetes,liverdisease,hypertension,other):
 		check_ci 	= (ci_user !=None) and (type(ci_user) == int)
 		check_other = type(other) == str
-		print("1")
+		#print("1")
 
 		if (check_ci and check_other):
-			print("2")	
+			#print("2")	
 			check_long_ci = 1 <= ci_user <= CONST_MAX_CI
 			check_long_other = CONST_MIN <= len(other) <= CONST_MAX_500
 			
 			if check_long_ci and check_long_other:
-				print("3")
+				#print("3")
 				old = FamilyBackground.query.filter_by(ci_user=ci_user).first()
 
 				if old == None:
-					print("4")
+					#print("4")
 					new = FamilyBackground(ci_user,asthma,
 										cancer,
 										heartdisease,
@@ -43,19 +43,19 @@ class familyBackground():
 	def updateFamilyBackground(self,ci_user,asthma,cancer,heartdisease,diabetes,liverdisease,hypertension,other):
 
 		check_ci 	= (ci_user !=None) and (type(ci_user) == int)
-		check_other = type(other) == str
-		print("1")
+		check_other = type(other) == str 
+		#print("1")
 		if (check_ci and check_other):
-			print("2")
+			#print("2")
 			check_long_ci = 1 <= ci_user <= CONST_MAX_CI
 			check_long_other = CONST_MIN <= len(other) <= CONST_MAX_500
 
 			if check_long_ci and check_long_other:
-				print("3")
+				#print("3")
 				old = FamilyBackground.query.filter_by(ci_user=ci_user).first()
 
 				if old != None:
-					print("4")
+					#print("4")
 					old.asthma=asthma
 					old.cancer=cancer
 					old.heartdisease=heartdisease
@@ -68,12 +68,16 @@ class familyBackground():
 		return False
 
 	def deleteFamilyBackground(self,ci_user):
-		
-		old = FamilyBackground.query.filter_by(ci_user=ci_user).first()
-		
-		if old != None:
-			db.session.delete(old)
-			db.session.commit()
-			return True
+		check_ci 	= (ci_user !=None) and (type(ci_user) == int)
+		if check_ci:
+			check_length = CONST_MIN < ci_user <= CONST_MAX_CI
+			
+			if check_length:
+				old = FamilyBackground.query.filter_by(ci_user=ci_user).first()
+				
+				if old != None:
+					db.session.delete(old)
+					db.session.commit()
+					return True
 
 		return False
